Fix hamburger menu ui issue
